3002.rose edit.traceback (#3003)
Fix traceback when trying to stop a previously stopped splash screen process.

+++ metomi/rose/gtk/splash.py
@@ -227,8 +227,9 @@ class SplashScreenProcess:
         )
 
     def stop(self):
-        self.process.kill()
-        self.process = None
+        if self.process:
+            self.process.kill()
+            self.process = None
 
 
 class SplashScreenUpdaterThread(threading.Thread):
